- ghettoproxy ( adding GPS protocols )WIP - fix BEARING_METHOD_3 - few fixes

+++ GhettoStation/MSP.cpp
@@ -135,7 +135,7 @@ void msp_check() {
   if (MSPcmd==MSP_ALTITUDE)
   {
     #ifdef BARO_ALT
-    uav_alt = read32() / 10;
+    uav_alt = (int16_t) (read32() / 10) ;
     #endif
     //uav_vario = read16();
   }
